hwtest/main: Fix CSP ID for ADCS System HK
Currently, we are likely facing an issue with the CAN Driver, which
is preventing us from enabling Routing on the MAIN Board and performing
the HW test. As a workaround, the MAIN Board is transmitting the
System HK of the ADCS Board instead. However, the CSP ID of the
System HK from the ADCS Board is set to that of the MAIN Board, so
this commit fixes this to the CSP ID of the ADCS Board.
This fix is only a workaround, and ultimately, these workarounds
will be unnecessary once the Routing issue is resolved.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>