lib/csp: Add file upload (OPEN) command
This commit adds the open file command needed for file upload from
the Ground System to SC-Sat1. This command includes the session ID
and the target file name. On the SC-Sat1 side, it associates the
session ID with the file name and opens the target file with write
permissions. If the target file does not exist, it will be newly
created. The session ID will be included in subsequent Data Packets
(DATA).

For uploads, the file that can be opened is defined by
CONFIG_SC_LIB_CSP_UPLOAD_MAX_SESSION, but please note that this must
be set to a value that is the same as or smaller than
CONFIG_ZVFS_OPEN_MAX.

This implementation also uses `slist` for session management; however,
since upload commands are processed sequentially from the CSP server
thread, mutual exclusion is not implemented at this time.

In a subsequent commit, the DATA and CLOSE commands will also be
added.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>