lib/csp: Add file upload (WRQ) command
This commit adds the Write Request (WRQ) command needed for file
upload from the Ground System to SC-Sat1. This command includes the
session ID and the target file name. On the SC-Sat1 side, it
associates the session ID with the file name and opens the target
file with write permissions. If the target file does not exist, it
will be newly created. The session ID will be included in subsequent
Data Packets (DATA).

SC-Sat1 will not reject a WRQ with a different file name for an
already-managed session ID. This approach is based on the current
assumption that the Ground System and SC-Sat1 maintain a
one-to-one relationship, simplifying the implementation.

BTW, The names `WRQ` and `DATA` are based on TFTP.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>