lib/csp: Add Upload (DATA) command
This commit adds the write data command needed for file upload from
the Ground System to SC-Sat1. This command includes the session ID
indicating the file opened with OPEN command, the write offset, the
size, and the actual datae.

The maximum length of the data to be sent can be specified in Kconfig.
Currently, it is set to 200 bytes, considering the uplink frame limit
of the SRS-3.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>